navbar: move user_is_conduit_admin helper to lando.main.auth (Bug 2042232)
