+++ src/lando/jinja.py
@@ -11,10 +11,10 @@ from django.utils.html import escape
 from jinja2 import Environment
 from markupsafe import Markup
 
+from lando.main.auth import user_is_conduit_admin
 from lando.main.models import JobStatus, LandingJob, Repo, UpliftJob
 from lando.main.models.revision import Revision
 from lando.main.scm import SCMType
-from lando.middleware import user_is_conduit_admin
 from lando.treestatus.models import (
     ReasonCategory,
     TreeCategory,
