+++ src/lando/middleware.py
@@ -7,13 +7,13 @@ from typing import Optional
 
 from django.conf import settings
 from django.contrib.auth import authenticate
-from django.contrib.auth.models import User
 from django.core.handlers.wsgi import WSGIRequest
 from django.http import HttpRequest, HttpResponse
 from django.template.loader import render_to_string
 from django.template.response import TemplateResponse
 from django.urls import resolve
 
+from lando.main.auth import user_is_conduit_admin
 from lando.main.models import ConfigurationKey, ConfigurationVariable
 from lando.utils.phabricator import (
     PhabricatorAPIException,
@@ -21,20 +21,6 @@ from lando.utils.phabricator import (
 )
 
 logger = logging.getLogger(__name__)
-
-CONDUIT_ADMIN_GROUP_NAME = "conduit-admin"
-
-
-def user_is_conduit_admin(user: User) -> bool:
-    """Return whether `user` is a Conduit administrator.
-
-    Administrators are staff users who belong to the `conduit-admin` group.
-    """
-    return (
-        user.is_authenticated
-        and user.is_staff
-        and user.groups.filter(name=CONDUIT_ADMIN_GROUP_NAME).exists()
-    )
 
 
 class ResponseHeadersMiddleware:
