+++ src/lando/tests/test_middlewares.py
@@ -1,8 +1,5 @@
 import pytest
-from django.contrib.auth.models import AnonymousUser, Group, User
 from django.test.client import Client
-
-from lando.middleware import CONDUIT_ADMIN_GROUP_NAME, user_is_conduit_admin
 
 
 @pytest.mark.parametrize(
@@ -37,32 +34,3 @@ def test_cors_acao_header(
         assert "access-control-allow-origin" not in resp.headers, (
             f"Unexpected ACAO header present for request from '{origin}' to '{path}'"
         )
-
-
-@pytest.mark.parametrize(
-    "is_staff,in_group,expected",
-    (
-        (True, True, True),
-        (True, False, False),
-        (False, True, False),
-        (False, False, False),
-    ),
-)
-@pytest.mark.django_db
-def test_user_is_conduit_admin(is_staff: bool, in_group: bool, expected: bool):
-    user = User.objects.create_user("testuser", is_staff=is_staff)
-    if in_group:
-        group, _ = Group.objects.get_or_create(name=CONDUIT_ADMIN_GROUP_NAME)
-        group.user_set.add(user)
-
-    assert user_is_conduit_admin(user) is expected, (
-        f"A staff={is_staff}, in_group={in_group} user should "
-        f"{'' if expected else 'not '}be a Conduit admin."
-    )
-
-
-@pytest.mark.django_db
-def test_user_is_conduit_admin_anonymous():
-    assert user_is_conduit_admin(AnonymousUser()) is False, (
-        "An unauthenticated user should not be a Conduit admin."
-    )
